+++ src/Renderer.OpenGL/Renderer/OpenGLRenderer.h
@@ -50,7 +50,7 @@ namespace Renderer
         virtual unsigned long requestVertexBuffer(const boost::shared_array<Utilities::Memory::byte>& data,
                                                   unsigned int numVertices, const Graphics::VertexFormat& fmt);
 
-        virtual unsigned long requestIndexBuffer(const boost::shared_array<unsigned short>& data, unsigned int numIndexes);
+        virtual unsigned long requestIndexBuffer(const boost::shared_array<unsigned int>& data, unsigned int numIndexes);
 
         virtual unsigned long requestEffect(const Utilities::IO::File& file);
         virtual unsigned long requestTexture(const Utilities::IO::File& file);
